do not align vararg args (indent normally)
Even with AlignAfterOpenBracket = DontAlign
a method call with varargs was aligning with the first space in the brackets like this:

NSString *s = [NSString stringWithFormat:@"hello %@",
                        name];

This was first introduced in commit 3aa9a6a1,
because AlignAfterOpenBracket was only applied if Current.FakeLParens precedence is prec::Comma.

The tests in that commit don't really explain why this check is needed,
but in this case the precendence in the square brackets is prec::Unknown,
so AlignAfterOpenBracket was not taken into account.
When the check for precedence is removed the DontAlign setting starts to work.

After:
NSString *s = [NSString stringWithFormat:@"hello %@",
    name];